pyproject.toml
[project]
name = "task-04"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "litellm>=1.80.0",
    "openai-agents>=0.6.1",
    "pypdf>=6.3.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
]
